Crate oxc_parser

Oxc Parser for JavaScript and TypeScript

Performance

The following optimization techniques are used:

AST is allocated in a memory arena (bumpalo) for fast AST drop

Short strings are inlined by CompactString

No other heap allocations are done except the above two

[oxc_span::Span] offsets uses u32 instead of usize

Scope binding, symbol resolution and complicated syntax errors are not done in the parser, they are delegated to the semantic analyzer

Conformance

The parser parses all of Test262 and most of Babel and TypeScript parser conformance tests.

See oxc coverage for details

Test262 Summary:
AST Parsed     : 44000/44000 (100.00%)

Babel Summary:
AST Parsed     : 2065/2071 (99.71%)

TypeScript Summary:
AST Parsed     : 2337/2337 (100.00%)

Usage

The parser has a minimal API with three inputs and one return struct (ParserReturn).

let parser_return = Parser::new(&allocator, &source_text, source_type).parse();

Visitor

See [oxc_ast::Visit] and [oxc_ast::VisitMut]

Visiting without a visitor

For ad-hoc tasks, the semantic analyzer can be used to get a parent pointing tree with untyped nodes, the nodes can be iterated through a sequential loop.

for node in semantic.nodes().iter() {
    match node.kind() {
        // check node
    }
}

See full linter example

Structs

Parser

Recursive Descent Parser for ECMAScript and TypeScript

ParserReturn

Return value of parser consisting of AST, errors and comments

Enums

Kind

Constants

MAX_LEN

Maximum length of source which can be parsed (in bytes). ~4 GiB on 64-bit systems, ~2 GiB on 32-bit systems.