Struct oxc_parser::Parser

pub struct Parser<'a> { /* private fields */ }

Implementations

impl<'a> Parser<'a>

pub const fn start_span(&self) -> Span

pub const fn end_span(&self, span: Span) -> Span

pub const fn cur_token(&self) -> &Token

Get current token

pub const fn cur_kind(&self) -> Kind

Get current Kind

pub fn cur_src(&self) -> &'a str

Get current source text

pub const fn cur_atom(&self) -> Option<&Atom>

Get current atom

pub fn peek_token(&mut self) -> &Token

Peek next token, returns EOF for final peek

pub fn peek_kind(&mut self) -> Kind

Peek next kind, returns EOF for final peek

pub fn peek_at(&mut self, kind: Kind) -> bool

Peek at kind

pub fn nth(&mut self, n: u8) -> &Token

Peek nth token

pub fn nth_at(&mut self, n: u8, kind: Kind) -> bool

Peek at nth kind

pub fn nth_kind(&mut self, n: u8) -> Kind

Peek nth kind

pub fn at(&self, kind: Kind) -> bool

Checks if the current index has token Kind

pub fn eat(&mut self, kind: Kind) -> bool

Advance and return true if we are at Kind, return false otherwise

pub fn bump(&mut self, kind: Kind)

Advance and return true if we are at Kind

pub fn bump_any(&mut self)

Advance any token

pub fn bump_remap(&mut self, kind: Kind)

Advance and change token type, useful for changing keyword to ident

pub fn asi(&mut self) -> Result<()>

Automatic Semicolon Insertion https://tc39.es/ecma262/#sec-automatic-semicolon-insertion

Errors

pub fn can_insert_semicolon(&self) -> bool

pub fn expect(&mut self, kind: Kind) -> Result<()>

Expect a Kind or return error

Errors

pub const fn current_range(&self) -> Span

pub fn expect_jsx_child(&mut self, kind: Kind) -> Result<()>

Expect the next next token to be a JsxChild, i.e. < or { or JSXText

Errors

pub fn expect_jsx_attribute_value(&mut self, kind: Kind) -> Result<()>

Expect the next next token to be a JsxString or any other token

Errors

pub fn read_regex(&mut self)

Tell lexer to read a regex

pub fn re_lex_template_substitution_tail(&mut self)

Tell lexer to read a template substitution tail

pub fn re_lex_jsx_identifier(&mut self)

Tell lexer to re-read a jsx identifier

pub fn re_lex_right_angle(&mut self) -> Kind

pub fn re_lex_ts_l_angle(&mut self)

pub fn re_lex_ts_r_angle(&mut self)

pub fn checkpoint(&self) -> ParserCheckpoint<'a>

pub fn rewind(&mut self, checkpoint: ParserCheckpoint<'a>)

pub fn try_parse<T>(
&mut self,
func: impl FnOnce(&mut Parser<'a>) -> Result<T>
) -> Result<T>

Errors

pub fn lookahead<U>(&mut self, predicate: impl Fn(&mut Parser<'a>) -> U) -> U

pub fn without_context<F, T>(&mut self, flags: Context, cb: F) -> Twhere
F: FnOnce(&mut Self) -> T,

pub fn with_context<F, T>(&mut self, flags: Context, cb: F) -> Twhere
F: FnOnce(&mut Self) -> T,

impl<'a> Parser<'a>

pub fn parse_binding_pattern(&mut self) -> Result<(BindingPattern<'a>, bool)>

Destructuring Binding Patterns LexicalBinding BindingIdentifier Initializer_opt BindingPattern Initializer BindingPattern: ObjectBindingPattern ArrayBindingPattern

pub fn parse_rest_element(&mut self) -> Result<Box<'a, RestElement<'a>>>

Section 14.3.3 Binding Rest Property

pub fn parse_binding_element(&mut self) -> Result<BindingPattern<'a>>

BindingElement SingleNameBinding BindingPattern Initializer

pub fn parse_object_pattern_property(&mut self) -> Result<Property<'a>>

impl<'a> Parser<'a>

Section 15.7 Class Definitions

pub fn parse_class_statement(
&mut self,
stmt_ctx: StatementContext
) -> Result<Statement<'a>>

pub fn parse_class_declaration(
&mut self,
declare: bool
) -> Result<Box<'a, Class<'a>>>

Section 15.7 Class Definitions

pub fn parse_class_expression(&mut self) -> Result<Expression<'a>>

Section Class Definitions https://tc39.es/ecma262/#prod-ClassExpression ClassExpression[Yield, Await] : class BindingIdentifier[?Yield, ?Await]opt ClassTail[?Yield, ?Await]

pub fn parse_heritage_clause(
&mut self
) -> Result<(Option<Vec<'a, (Expression<'a>, Option<Box<'a, TSTypeParameterInstantiation<'a>>>, Span)>>, Option<Vec<'a, Box<'a, TSClassImplements<'a>>>>)>

pub fn parse_class_element(&mut self) -> Result<ClassElement<'a>>

impl<'a> Parser<'a>

pub fn parse_let(&mut self, stmt_ctx: StatementContext) -> Result<Statement<'a>>

pub fn parse_variable_declaration(
&mut self,
decl_ctx: VariableDeclarationContext
) -> Result<Box<'a, VariableDeclaration<'a>>>

impl<'a> Parser<'a>

pub fn parse_paren_expression(&mut self) -> Result<Expression<'a>>

pub fn parse_expression(&mut self) -> Result<Expression<'a>>

Section Expression https://tc39.es/ecma262/#sec-ecmascript-language-expressions

pub fn parse_identifier_expression(&mut self) -> Result<Expression<'a>>

PrimaryExpression: Identifier Reference

pub fn parse_identifier_reference(&mut self) -> Result<IdentifierReference>

pub fn parse_binding_identifier(&mut self) -> Result<BindingIdentifier>

BindingIdentifier : Identifier

pub fn parse_label_identifier(&mut self) -> Result<LabelIdentifier>

pub fn parse_identifier_name(&mut self) -> Result<IdentifierName>

pub fn parse_keyword_identifier(&mut self, kind: Kind) -> IdentifierName

Parse keyword kind as identifier

pub fn parse_identifier_kind(&mut self, kind: Kind) -> (Span, Atom)

pub fn parse_private_identifier(&mut self) -> PrivateIdentifier

Section https://tc39.es/ecma262/#prod-PrivateIdentifier PrivateIdentifier :: # IdentifierName

Panics

pub fn parse_literal_expression(&mut self) -> Result<Expression<'a>>

Literal Expression parses string | true | false | null | number

pub fn parse_literal_boolean(&mut self) -> Result<BooleanLiteral>

pub fn parse_literal_null(&mut self) -> NullLiteral

pub fn parse_literal_number(&mut self) -> Result<NumberLiteral<'a>>

pub fn parse_literal_bigint(&mut self) -> Result<BigintLiteral>

pub fn parse_literal_regexp(&mut self) -> Result<RegExpLiteral>

pub fn parse_literal_string(&mut self) -> Result<StringLiteral>

pub fn parse_array_expression(&mut self) -> Result<Expression<'a>>

Section Array Expression https://tc39.es/ecma262/#prod-ArrayLiteral ArrayLiteral[Yield, Await]: [ Elision opt ] [ ElementList[?Yield, ?Await] ] [ ElementList[?Yield, ?Await] , Elisionopt ]

pub fn parse_template_element(&mut self, tagged: bool) -> TemplateElement

pub fn parse_lhs_expression(&mut self) -> Result<Expression<'a>>

Section 13.3 Left-Hand-Side Expression

pub fn parse_unary_expression_base(
&mut self,
lhs_span: Span
) -> Result<Expression<'a>>

Section 13.5 Unary Expression

pub fn parse_assignment_expression_base(&mut self) -> Result<Expression<'a>>

pub fn parse_assignment_expression(&mut self) -> Result<Expression<'a>>

AssignmentExpression[In, Yield, Await] :

impl<'a> Parser<'a>

pub fn at_function_with_async(&mut self) -> bool

pub fn at_async_no_new_line(&mut self) -> bool

pub fn parse_function_body(&mut self) -> Result<Box<'a, FunctionBody<'a>>>

pub fn parse_formal_parameters(
&mut self,
params_kind: FormalParameterKind
) -> Result<Box<'a, FormalParameters<'a>>>

pub fn parse_function(
&mut self,
span: Span,
id: Option<BindingIdentifier>,
async: bool,
generator: bool,
func_kind: FunctionKind
) -> Result<Box<'a, Function<'a>>>

pub fn parse_function_declaration(
&mut self,
stmt_ctx: StatementContext
) -> Result<Statement<'a>>

Function Declaration

pub fn parse_function_impl(
&mut self,
func_kind: FunctionKind
) -> Result<Box<'a, Function<'a>>>

pub fn parse_function_expression(
&mut self,
span: Span,
async: bool
) -> Result<Expression<'a>>

Function Expression

pub fn parse_single_param_function_expression(
&mut self,
span: Span,
async: bool,
generator: bool
) -> Result<Expression<'a>>

pub fn parse_method(
&mut self,
async: bool,
generator: bool
) -> Result<Box<'a, Function<'a>>>

Section 15.4 Method Definitions ClassElementName ( UniqueFormalParameters ) { FunctionBody } GeneratorMethod

ClassElementName AsyncMethod async ClassElementName AsyncGeneratorMethod async * ClassElementName

pub fn parse_yield_expression(&mut self) -> Result<Expression<'a>>

Section 15.5 Yield Expression yield yield [no LineTerminator here] AssignmentExpression yield [no LineTerminator here] * AssignmentExpression

pub fn parse_function_id(
&mut self,
kind: FunctionKind,
async: bool,
generator: bool
) -> Option<BindingIdentifier>

pub fn is_parenthesized_arrow_function_expression(
&mut self,
async: bool
) -> IsParenthesizedArrowFunction

pub fn is_parenthesized_arrow_function(
&mut self
) -> IsParenthesizedArrowFunction

pub fn parse_parenthesized_arrow_function_head(
&mut self
) -> Result<(Option<Box<'a, TSTypeParameterDeclaration<'a>>>, Box<'a, FormalParameters<'a>>, Option<TSTypeAnnotation<'a>>, bool, Span)>

pub fn parse_arrow_function_body(
&mut self,
span: Span,
type_parameters: Option<Box<'a, TSTypeParameterDeclaration<'a>>>,
params: Box<'a, FormalParameters<'a>>,
return_type: Option<TSTypeAnnotation<'a>>,
async: bool
) -> Result<Expression<'a>>

ConciseBody [lookahead ≠ {] ExpressionBody[?In, ~Await] { FunctionBody[~Yield, ~Await] } ExpressionBody[In, Await] : AssignmentExpression[?In, ~Yield, ?Await]

pub fn parse_parenthesized_arrow_function(&mut self) -> Result<Expression<'a>>

Section Arrow Function https://tc39.es/ecma262/#sec-arrow-function-definitions ArrowFunction[In, Yield, Await] : ArrowParameters[?Yield, ?Await] [no LineTerminator here] => ConciseBody[?In]

impl<'a> Parser<'a>

pub fn parse_import_expression(&mut self, span: Span) -> Result<Expression<'a>>

Import Call ImportCall : import ( AssignmentExpression )

pub fn parse_import_declaration(&mut self) -> Result<Statement<'a>>

Section 16.2.2 Import Declaration

pub fn parse_ts_export_assignment_declaration(
&mut self
) -> Result<Box<'a, TSExportAssignment<'a>>>

pub fn parse_ts_export_namespace(
&mut self
) -> Result<Box<'a, TSNamespaceExportDeclaration>>

pub fn parse_export_declaration(&mut self) -> Result<Statement<'a>>

Exports https://tc39.es/ecma262/#sec-exports

pub fn parse_module_export_name(&mut self) -> Result<ModuleExportName>

impl<'a> Parser<'a>

pub fn parse_object_expression(&mut self) -> Result<Expression<'a>>

Object Expression ObjectLiteral[Yield, Await] : { } { PropertyDefinitionList[?Yield, ?Await] } { PropertyDefinitionList[?Yield, ?Await] , }

pub fn parse_property_definition(&mut self) -> Result<Box<'a, Property<'a>>>

PropertyDefinition[Yield, Await]

pub fn parse_spread_element(&mut self) -> Result<Box<'a, SpreadElement<'a>>>

PropertyDefinition[Yield, Await] : … AssignmentExpression[+In, ?Yield, ?Await]

pub fn parse_property_name(&mut self) -> Result<(PropertyKey<'a>, bool)>

PropertyName[Yield, Await] : LiteralPropertyName ComputedPropertyName[?Yield, ?Await]

pub fn parse_computed_property_name(&mut self) -> Result<Expression<'a>>

ComputedPropertyName[Yield, Await] : [ AssignmentExpression[+In, ?Yield, ?Await] ]

impl<'a> Parser<'a>

pub fn parse_directives_and_statements(
&mut self,
is_top_level: bool
) -> Result<(Vec<'a, Directive<'a>>, Vec<'a, Statement<'a>>)>

https://tc39.es/ecma262/#prod-StatementList StatementList[Yield, Await, Return] : StatementListItem[?Yield, ?Await, ?Return] StatementList[?Yield, ?Await, ?Return] StatementListItem[?Yield, ?Await, ?Return]

pub fn parse_statement_list_item(
&mut self,
stmt_ctx: StatementContext
) -> Result<Statement<'a>>

StatementListItem[Yield, Await, Return] : Statement[?Yield, ?Await, ?Return] Declaration[?Yield, ?Await]

pub fn parse_block(&mut self) -> Result<Box<'a, BlockStatement<'a>>>

Section 14.2 Block Statement

pub fn parse_block_statement(&mut self) -> Result<Statement<'a>>

pub fn parse_variable_statement(
&mut self,
stmt_ctx: StatementContext
) -> Result<Statement<'a>>

Section 14.3.2 Variable Statement

pub fn parse_expression_statement(
&mut self,
span: Span,
expression: Expression<'a>
) -> Result<Statement<'a>>

Section 14.5 Expression Statement

pub fn parse_switch_case(&mut self) -> Result<SwitchCase<'a>>

impl<'a> Parser<'a>

pub fn parse_jsx_expression(&mut self) -> Result<Expression<'a>>

impl<'a> Parser<'a>

pub fn at_start_of_ts_declaration(&mut self) -> bool

impl<'a> Parser<'a>

pub fn is_at_enum_declaration(&mut self) -> bool

—————–– Enum ——————

pub fn parse_ts_enum_declaration(
&mut self,
declare: bool,
span: Span
) -> Result<Declaration<'a>>

https://www.typescriptlang.org/docs/handbook/enums.html

pub fn parse_ts_enum_member(&mut self) -> Result<TSEnumMember<'a>>

pub fn parse_ts_type_annotation(
&mut self
) -> Result<Option<TSTypeAnnotation<'a>>>

—————–– Annotation —————–

pub fn parse_ts_variable_annotation(
&mut self
) -> Result<(Option<TSTypeAnnotation<'a>>, bool)>

pub fn parse_ts_type_alias_declaration(
&mut self,
declare: bool,
span: Span
) -> Result<Declaration<'a>>

pub fn parse_ts_interface_declaration(
&mut self,
declare: bool,
span: Span
) -> Result<Declaration<'a>>

——————— Interface ————————

pub fn is_at_interface_declaration(&mut self) -> bool

pub fn parse_ts_type_signature(&mut self) -> Result<TSSignature<'a>>

pub fn is_at_ts_index_signature_member(&mut self) -> bool

Must be at [ident: or <modifiers> [ident:

pub fn is_nth_at_modifier(
&mut self,
n: u8,
is_constructor_parameter: bool
) -> bool

pub fn parse_ts_namespace_or_module_declaration_body(
&mut self,
span: Span,
declare: bool
) -> Result<Box<'a, TSModuleDeclaration<'a>>>

pub fn parse_ts_namespace_or_module_declaration(
&mut self,
declare: bool
) -> Result<Box<'a, TSModuleDeclaration<'a>>>

pub fn parse_ts_global_declaration(
&mut self
) -> Result<Box<'a, TSModuleDeclaration<'a>>>

pub fn parse_ts_namespace_or_module_statement(
&mut self,
declare: bool
) -> Result<Statement<'a>>

pub fn parse_ts_global_statement(&mut self) -> Result<Statement<'a>>

pub fn is_nth_at_ts_namespace_declaration(&mut self, n: u8) -> bool

pub fn is_at_ts_declare_clause(&mut self) -> bool

———————– declare ———————

pub fn parse_ts_declaration_statement(&mut self) -> Result<Statement<'a>>

pub fn parse_declaration_clause(&mut self) -> Result<Declaration<'a>>

pub fn parse_ts_declare_function(&mut self) -> Result<Box<'a, Function<'a>>>

pub fn parse_ts_type_assertion(&mut self) -> Result<Expression<'a>>

pub fn parse_ts_import_equals_declaration(
&mut self,
span: Span,
is_export: bool
) -> Result<Declaration<'a>>

pub fn parse_ts_this_parameter(&mut self) -> Result<()>

pub fn eat_decorators(&mut self) -> Result<()>

impl<'a> Parser<'a>

pub fn parse_ts_type(&mut self) -> Result<TSType<'a>>

pub fn parse_ts_type_parameters(
&mut self
) -> Result<Option<Box<'a, TSTypeParameterDeclaration<'a>>>>

pub fn parse_ts_implements_clause(
&mut self
) -> Result<Vec<'a, Box<'a, TSClassImplements<'a>>>>

pub fn parse_ts_type_parameter(
&mut self
) -> Result<Box<'a, TSTypeParameter<'a>>>

pub fn parse_ts_qualified_name(&mut self) -> Result<TSTypeName<'a>>

pub fn parse_ts_type_arguments(
&mut self
) -> Result<Option<Box<'a, TSTypeParameterInstantiation<'a>>>>

pub fn parse_ts_type_arguments_in_expression(
&mut self
) -> Result<Option<Box<'a, TSTypeParameterInstantiation<'a>>>>

pub fn is_at_named_tuple_element(&mut self) -> bool

pub fn parse_ts_return_type_annotation(
&mut self
) -> Result<Option<TSTypeAnnotation<'a>>>

pub fn parse_ts_return_type(&mut self) -> Result<TSType<'a>>

pub fn is_next_at_type_member_name(&mut self) -> bool

pub fn parse_ts_call_signature_member(&mut self) -> Result<TSSignature<'a>>

pub fn parse_ts_getter_signature_member(&mut self) -> Result<TSSignature<'a>>

pub fn parse_ts_setter_signature_member(&mut self) -> Result<TSSignature<'a>>

pub fn parse_ts_property_or_method_signature_member(
&mut self
) -> Result<TSSignature<'a>>

pub fn parse_ts_constructor_signature_member(
&mut self
) -> Result<TSSignature<'a>>

pub fn parse_ts_index_signature_member(&mut self) -> Result<TSSignature<'a>>

pub fn parse_class_element_modifiers(
&mut self,
is_constructor_parameter: bool
) -> ModifierFlags

impl<'a> Parser<'a>

pub fn new(
allocator: &'a Allocator,
source: &'a str,
source_type: SourceType
) -> Self

pub fn allow_return_outside_function(self, allow: bool) -> Self

pub fn parse(self) -> ParserReturn<'a>

Parser main entry point Returns an empty Program on unrecoverable error, Recoverable errors are stored inside errors.